+++ eventory-server/src/main/java/com/eventory/expoAdmin/controller/ExpoAdminController.java
@@ -5,6 +5,7 @@ import com.eventory.common.exception.CustomErrorCode;
 import com.eventory.common.exception.CustomException;
 import com.eventory.expoAdmin.dto.*;
 import com.eventory.expoAdmin.service.ExpoAdminService;
+import com.eventory.expoAdmin.service.SalesAdminService;
 import org.springframework.core.io.Resource;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
@@ -22,6 +23,7 @@ import java.util.Map;
 public class ExpoAdminController {
 
     private final ExpoAdminService expoAdminService;
+    private final SalesAdminService salesAdminService;
 
     // 해당 박람회 관리자에 속하는 전체 박람회 목록
     @GetMapping("/expos")
@@ -34,7 +36,7 @@ public class ExpoAdminController {
     // 누적 매출, 총 결제 건수, 총 환불 건수
     @GetMapping("/expos/{expoId}/sales")
     public ResponseEntity<SalesResponseDto> findSalesStatistics(@PathVariable Long expoId) {
-        SalesResponseDto salesResponseDto = expoAdminService.findSalesStatistics(expoId);
+        SalesResponseDto salesResponseDto = salesAdminService.findSalesStatistics(expoId);
         return ResponseEntity.ok(salesResponseDto);
     }
 
@@ -42,9 +44,9 @@ public class ExpoAdminController {
     @GetMapping("/expos/{expoId}/stats")
     public ResponseEntity<List<Map<String, Object>>> findSales(@PathVariable Long expoId, @RequestParam String range) {
         List<Map<String, Object>> sales = switch (range.toLowerCase()) {
-            case "daily" -> expoAdminService.findDailySales(expoId);
-            case "monthly" -> expoAdminService.findMonthlySales(expoId);
-            case "yearly" -> expoAdminService.findYearlySales(expoId);
+            case "daily" -> salesAdminService.findDailySales(expoId);
+            case "monthly" -> salesAdminService.findMonthlySales(expoId);
+            case "yearly" -> salesAdminService.findYearlySales(expoId);
             default -> throw new CustomException(CustomErrorCode.NOT_FOUNT_RANGE);
         };
         return ResponseEntity.ok(sales);
@@ -61,9 +63,9 @@ public class ExpoAdminController {
         List<PaymentResponseDto> paymentResponseDto;
 
         if (page!=null && size!=null) { // 페이징 O
-            paymentResponseDto = expoAdminService.findAllPayments(expoId, code, page, size);
+            paymentResponseDto = salesAdminService.findAllPayments(expoId, code, page, size);
         } else { // 페이징 X
-            paymentResponseDto = expoAdminService.findAllPayments(expoId, code);
+            paymentResponseDto = salesAdminService.findAllPayments(expoId, code);
         }
 
         return ResponseEntity.ok(paymentResponseDto);
@@ -73,9 +75,9 @@ public class ExpoAdminController {
     @PostMapping("/expos/{expoId}/payment/report")
     public ResponseEntity<Resource> downloadPaymentsExcel(@PathVariable Long expoId) {
 
-        List<PaymentResponseDto> paymentResponseDto = expoAdminService.findAllPayments(expoId, null);
+        List<PaymentResponseDto> paymentResponseDto = salesAdminService.findAllPayments(expoId, null);
 
-        Resource excel = expoAdminService.downloadPaymentsExcel(paymentResponseDto);
+        Resource excel = salesAdminService.downloadPaymentsExcel(paymentResponseDto);
 
         return ResponseEntity.ok()
                 .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=payment.xlsx")
@@ -91,7 +93,7 @@ public class ExpoAdminController {
             @RequestParam(required = false, defaultValue = "0") Integer page,
             @RequestParam(required = false, defaultValue = "7") Integer size) {
 
-        List<RefundResponseDto> refundResponseDto = expoAdminService.findAllRefunds(expoId, status, page, size);
+        List<RefundResponseDto> refundResponseDto = salesAdminService.findAllRefunds(expoId, status, page, size);
 
         return ResponseEntity.ok(refundResponseDto);
     }
@@ -99,7 +101,7 @@ public class ExpoAdminController {
     // 환불 상태 변경
     @PatchMapping("/refund/{refundId}/status")
     public ResponseEntity<Void> updateRefundStatus(@PathVariable Long refundId, @RequestBody RefundRequestDto request) {
-        expoAdminService.updateRefundStatus(refundId, request);
+        salesAdminService.updateRefundStatus(refundId, request);
         return ResponseEntity.ok().build();
     }
 }
